+++ src/cli/commands/import.command.ts
@@ -1,7 +1,8 @@
 import { Command } from './command.interface.js';
 import { TSVFileReader } from '../../shared/libs/file-reader/index.js';
-import { OfferInterface } from '../../shared/types/offer.interface.js';
+
 import { getErrorMessage } from '../../shared/helpers/index.js';
+import { OfferInterface } from '../../shared/types/index.js';
 
 export class ImportCommand implements Command {
   private onImportedOffer(offer: OfferInterface): void {
