+++ src/shared/libs/file-reader/tsv-file-reader.ts
@@ -1,16 +1,16 @@
 import EventEmitter from 'node:events';
 import { createReadStream } from 'node:fs';
 import { FileReader } from './file-reader.interface.js';
-import { OfferInterface } from '../../types/offer.interface.js';
+import { OfferInterface } from '../../types/index.js';
 import {
   AmenitiesEnum,
   CityNameEnum,
   PropertyTypeEnum,
   UserTypeEnum,
 } from '../../types/enums.js';
-import { UserInterface } from '../../types/user.interface.js';
-import { CoordinatesType } from '../../types/coordinates.type.js';
-import { CITIES, CityType } from '../../types/city.type.js';
+import { UserInterface } from '../../types/index.js';
+import { CoordinatesType } from '../../types/index.js';
+import { CITIES, CityType } from '../../types/index.js';
 
 export class TSVFileReader extends EventEmitter implements FileReader {
   private CHUNK_SIZE = 16384; // 16KB
