+++ src/shared/types/index.ts
@@ -1,0 +1,7 @@
+export { CityType, CITIES } from './city.type.js';
+export { CommentInterface } from './comment.interface.js';
+export { CoordinatesType } from './coordinates.type.js';
+export { OfferInterface } from './offer.interface.js';
+export { UserInterface } from './user.interface.js';
+
+export { MockServerData } from './mock-server-data.type.js';
